ensure tests for .vbox_version and .vmfusion_version to avoid clashes

+++ templates/CentOS-6.1-x86_64-minimal/virtualbox.sh
@@ -1,8 +1,11 @@
-# Installing the virtualbox guest additions
-VBOX_VERSION=$(cat /home/veewee/.vbox_version)
-cd /tmp
-mount -o loop /home/veewee/VBoxGuestAdditions_$VBOX_VERSION.iso /mnt
-sh /mnt/VBoxLinuxAdditions.run
-umount /mnt
-rm -rf /home/veewee/VBoxGuestAdditions_*.iso
+if test -f .vbox_version ; then
 
+  # Installing the virtualbox guest additions
+  VBOX_VERSION=$(cat /home/veewee/.vbox_version)
+  cd /tmp
+  mount -o loop /home/veewee/VBoxGuestAdditions_$VBOX_VERSION.iso /mnt
+  sh /mnt/VBoxLinuxAdditions.run
+  umount /mnt
+  rm -rf /home/veewee/VBoxGuestAdditions_*.iso
+
+fi
