+++ templates/ubuntu-12.10-server-i386-packages/virtualbox.sh
@@ -3,13 +3,17 @@ then
   . .veewee_params
 fi
 
-# Installing the virtualbox guest additions
-apt-get -y install dkms
-VBOX_VERSION=$(cat /home/vagrant/.vbox_version)
-mount -o loop VBoxGuestAdditions_$VBOX_VERSION.iso /mnt
-sh /mnt/VBoxLinuxAdditions.run
-umount /mnt
+if test -f .vbox_version ; then
 
-rm VBoxGuestAdditions_$VBOX_VERSION.iso
+  # Installing the virtualbox guest additions
+  apt-get -y install dkms
+  VBOX_VERSION=$(cat /home/vagrant/.vbox_version)
+  mount -o loop VBoxGuestAdditions_$VBOX_VERSION.iso /mnt
+  sh /mnt/VBoxLinuxAdditions.run
+  umount /mnt
 
-cd /tmp
+  rm VBoxGuestAdditions_$VBOX_VERSION.iso
+  
+  cd /tmp
+
+fi
